requirements.txt
flask
python-dotenv
PyMuPDF
Pillow
requests
flask-cors
